feat: 店铺管理员

+++ app/Services/ShopManagerService.php
@@ -6,13 +6,14 @@ use App\Models\ShopManager;
 
 class ShopManagerService extends BaseService
 {
-    public function createManager($shopId, $userId)
+    public function createManager($shopId, $userId, $roleId)
     {
-        $address = ShopManager::new();
-        $address->shop_id = $shopId;
-        $address->user_id = $userId;
-        $address->save();
-        return $address;
+        $manager = ShopManager::new();
+        $manager->shop_id = $shopId;
+        $manager->user_id = $userId;
+        $manager->role_id = $roleId;
+        $manager->save();
+        return $manager;
     }
 
     public function getManagerList($shopId, $columns = ['*'])
@@ -20,8 +21,8 @@ class ShopManagerService extends BaseService
         return ShopManager::query()->where('shop_id', $shopId)->get($columns);
     }
 
-    public function deleteManager($shopId)
+    public function deleteManager($shopId, $userId)
     {
-        ShopManager::query()->where('shop_id', $shopId)->delete();
+        ShopManager::query()->where('shop_id', $shopId)->where('user_id', $userId)->delete();
     }
 }
